requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-render-block==0.9.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
mysqlclient==2.1.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
